Stage I. Engine with macros.
